+++ libs/meetings/__init__.py
@@ -1,0 +1,3 @@
+from .uid import UID_PREFIX, canonical_meeting_uid
+
+__all__ = ["UID_PREFIX", "canonical_meeting_uid"]